--- docs/neural_modeling/examples_utils/data.py
@@ -1,77 +1,0 @@
-#!/usr/bin/env python3
-
-from typing import Union
-
-import numpy as np
-import pynapple as nap
-
-TsdType = Union[nap.Tsd, nap.TsdFrame, nap.TsdTensor]
-
-import fsspec
-import h5py
-from dandi.dandiapi import DandiAPIClient
-from fsspec.implementations.cached import CachingFileSystem
-
-# Dandi stuffs
-from pynwb import NWBHDF5IO
-
-
-def download_dandi_data(dandiset_id, filepath):
-    with DandiAPIClient() as client:
-        asset = client.get_dandiset(dandiset_id, "draft").get_asset_by_path(filepath)
-        s3_url = asset.get_content_url(follow_redirects=1, strip_query=True)
-    
-    # first, create a virtual filesystem based on the http protocol
-    fs = fsspec.filesystem("http")
-
-    # create a cache to save downloaded data to disk (optional)
-    fs = CachingFileSystem(
-        fs=fs,
-        cache_storage="nwb-cache",  # Local folder for the cache
-    )
-
-    # next, open the file
-    file = h5py.File(fs.open(s3_url, "rb"))
-    io = NWBHDF5IO(file=file, load_namespaces=True)
-
-    return io
-
-
-def fill_forward(time_series, data, ep=None, out_of_range=np.nan):
-    """
-    Fill a time series forward in time with data.
-
-    Parameters
-    ----------
-    time_series:
-        The time series to match.
-    data: Tsd, TsdFrame, or TsdTensor
-        The time series with data to be extend.
-
-    Returns
-    -------
-    : Tsd, TsdFrame, or TsdTensor
-        The data time series filled forward.
-
-    """
-    assert isinstance(data, TsdType)
-
-    if ep is None:
-        ep = time_series.time_support
-    else:
-        assert isinstance(ep, nap.IntervalSet)
-        time_series.restrict(ep)
-
-    data = data.restrict(ep)
-    starts = ep.start
-    ends = ep.end
-
-    filled_d = np.full((time_series.t.shape[0], *data.shape[1:]), out_of_range, dtype=data.dtype)
-    fill_idx = 0
-    for start, end in zip(starts, ends):
-        data_ep = data.get(start, end)
-        ts_ep = time_series.get(start, end)
-        idxs = np.searchsorted(data_ep.t, ts_ep.t, side="right") - 1
-        filled_d[fill_idx:fill_idx + ts_ep.t.shape[0]][idxs >= 0] = data_ep.d[idxs[idxs>=0]]
-        fill_idx += ts_ep.t.shape[0]
-    return type(data)(t=time_series.t, d=filled_d, time_support=ep)
